+++ @ecomplus/storefront-twbs/src/index.js
@@ -9,7 +9,6 @@ import { name, version } from './../package.json'
 import $ from './$'
 import animateCss from './animate-css'
 import './bootstrap'
-import './fontawesome-icons'
 
 const _self = name + '@' + version