+++ @ecomplus/widget-offers-notification/src/index.js
@@ -1,0 +1,3 @@
+export default options => {
+  console.log('Offers notification widget started with EJS appends')
+}